Refactored LineShape and KappaMode
